fix(docs): move VersionTagFilter to docs/conf.py
Module imports are weird for docs since the project root changes. This
import was failing when building the docs in readthedocs:

Traceback (most recent call last):
  File "/home/docs/checkouts/readthedocs.org/user_builds/ddtrace/envs/latest/lib/python3.8/site-packages/sphinx/config.py", line 368, in eval_config_file
    execfile_(filename, namespace)
  File "/home/docs/checkouts/readthedocs.org/user_builds/ddtrace/envs/latest/lib/python3.8/site-packages/sphinx/util/pycompat.py", line 150, in execfile_
    exec_(code, _globals)
  File "/home/docs/checkouts/readthedocs.org/user_builds/ddtrace/checkouts/latest/docs/conf.py", line 22, in <module>
    from docs.ext.filters import VersionTagFilter
ModuleNotFoundError: No module named 'docs.ext'
